Move book model into appropriate folder
