+++ lib/app/data/firebase_api.dart
@@ -1,5 +1,5 @@
 import 'package:bmagrifa_books/app/data/base_api.dart';
-import 'package:bmagrifa_books/app/modules/book_info/books_model.dart';
+import 'package:bmagrifa_books/app/models/books_model.dart';
 import 'package:firebase_database/firebase_database.dart';
 
 class FirebaseApi implements BaseApi {
